Rename the data folder: "filesystem" -> "data"
Our lesson tempate says that data files should be in a "data" directory.
The name "filesystem" is not used in the lesson, so we are free to change it.
This also allows to create a script at the lesson-template level to zip this data
easily across different lessons and make them available throught the .getready callout.