Fix text error "home" to "current" directory
The text erroneously said "If she is in her home directory, Nelle can see what files she has using the command:", however, Nelle actually needs to be in the `data-shell` directory (set as current directory in the previous command) for this to work. 
Error pointed out by  =@=JohnRMoreau in PR /issues/314 discussion.