Update 07-find.md
Making a note about the '.gitkeep' file. Since this shows up in the output of 'find . -f', as can be seen on line 287
(not added by this pull request, there already), and since the 'writing/old' directory isn't used elsewhere in the
lesson, maybe we could delete that directory? This would make the diagram correct as it is....