Separate shell grammar from filesystem navigation
Episode 2 intertwines explaination of concepts and commands for
navigating the filesystem with asides about options, commands, spaces,
and so on. The "grammar" of what a shell command looks like is a
distinct topic from the concepts and commands for navigation, so I've
pulled the grammar parts into Episode 1.

I think this makes Episode 2 clearer, and provides a conceptual model
for commands before introducing them

In the process I've described the shell a little less pessimistically
in Episode one