Removed 'ls is alphabetical by default'
This is only the case with the GNU implementation - the BSD version of ls (which is the default on OSX, for example) does not list files alphabetically.  The POSIX standard explicitly says the output formatting for terminals and the ordering is implementation specific (non-terminal output is guaranteed to be one file per line).

This change therefore addresses both the misinformation that ls always lists alphabetically by default and improves POSIX compliance of the text (for which there is also an outstanding issue).

I have also changed the word 'terminal' to 'settings', where is says the output may or may not be coloured by default - again this it determined by more than just the terminal supporting and having colour enabled (most importantly by where 'ls' has be aliased to a default set of flags including the colour one - RedHat do this but Debian do not, for example) so to say it is just the terminal is misleading - I believe substituting this word addresses the inaccuracy without reducing understandability or increasing the complexity of the material.

Fixes /issues/744.  References /issues/708.