05-Loops: Fixing keyboard commands and adding examples to history usage (/issues/1078)
* Fixing capitalization of Ctrl-r and adding more detailed explanation

Ctrl-R doesn't work on all OS's while Ctrl-r does and is less ambiguous.

Same for Ctrl-C, using Ctrl-c is clearer

I also added an extra example of using the history command

* Fixing keyboard commands and adding history examples

I fixed the highlighting of keyboard commands to comply with the style guide.
These included <kbd>Ctrl</kbd>+<kbd>R</kbd> for example, as well as Ctrl-C and
Enter in a few places.

I also added some explanation to the history lookup commands and provided an
additional mechanism of using !head:p to recall the last command staring with 'head'.

* Removed a section I added at Gabriel's request

Gabriel thought the section I added went beyond the scope of an
introductory lesson.  I've removed it to get the rest of the PR resolved.