Edited wildcard exercise
The existing proposed solution to files ending in A.txt or B.txt exercise at the end says it uses two expressions, but actually it uses two commands. However the two command solution is useful for the follow up questions. I reworded the the wildcard exercise at the end to use the term commands instead of expressions, and also included an extra solution with two expressions (arguments)