Update 03-create.md
Added lines 77 - 99 to introduce mkdir -p, along with a sample, as well as the introduction of the -R flag to ls to recursively list the newly created directory hierarchy beneath the thesis directory.  Added lines 877 - 879 to introduce an exercise scenario to highlight mkdir -p.  Added "the default behavior of" to line 894 to avoid leaving students with the impression that mkdir cannot create subdirectories.  Added lines 897-899 to provide a solution explanation for the mkdir -p exercise scenario.