01-intro.md: mention nontrivial prompts
A lot of standard environments set a prompt that is not just a dollar
sign. However, only the latter is used in the lecture. This can create
confusion for learners, as their environment most often does not look
like the one of the lecturer. This commonly happens because user or host
names are included in the prompt by default and those most often are
different not only between lecturer and learner, but also between
learners.

This commit adds a short paragraph that explains that what people see
before the $-sign might be different and include things like their user
name. It also gives a simple example, before it specifically mentiones
that regardless of what is before that dollar sign: they should not worry.
Finally, it again puts emphasis on the dollar sign itself as the only
thing they should focus on, because this will be used later in the
course. It leaves the details to those places, as this is too early to
go into these things without creating even more confusion.