05-loop.md: mention the now additional `echo`
The pull request =/=swcarpentry/=/=shell-novice/pull/1274
added an echo to have an easier time following along what value the
variable holds at any time in the loop. This additional commit to
that pull request adds the mention of that echo in the text below,
as requested in =/=swcarpentry/=/=shell-novice/pull/1274#issuecomment-1119218771
by https://github.com/deppen8.